partsum as a benchmark of CPU superscalarity/branch prediction capabilities
It's relatively easy to measure the effect of the compiler optimizations.
One compares the run time of the optimized (compiled with `-O2`) versus
the non-optimized (compiled with `-O0`) builds. The faster the optimized
build runs (relatively to non-optimized one) the better the compiler
optimizations are.

It would be nice to measure the effect of the **hardware** optimizations
(pipelining, superscalar execution, out-of-order execution, speculative
execution, etc). However CPUs don't have `-O0` flag. The trick is to
1) disable the **software** (compiler) optimizations, and
2) randomize the program data so branch predictions are useless

Put it another way `rand_partsum` is kind of `partsum` with some
hardware optimizations disabled. Hence comparing the run time of
`partsum` (hardware optimized code) with `rand_partum` (the same
code with some hardware optimizations disabled) tells how good
the hardware optimizations are.